backend/requirements.txt
Django==3.2.3
django-cors-headers==3.13.0
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.2.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
reportlab==3.6.11